Reduce compilation memory use - move MergeUpdateClause and ResampleClause in to their own files
clause.cpp was the heaviest non-pybind11 TU at 3,924 MB baseline. Splitting MergeUpdateClause and ResampleClause into their own files reduced it to 3,381 MB (-543 MB, -14%). The two new TUs compile at 2,362 MB and 1,806 MB respectively.

Remove async_store.hpp from versioned_engine.hpp

This stops version_store_api.cpp from pulling in clause.hpp and the processing pipeline machinery. It reduced its compilation memory use from 3362MB to 2459MB.
