+++ cpp/arcticdb/CMakeLists.txt
@@ -511,6 +511,8 @@ set(arcticdb_srcs
         processing/processing_unit.cpp
         processing/aggregation_utils.cpp
         processing/clause.cpp
+        processing/clause_merge_update.cpp
+        processing/clause_resample.cpp
         processing/clause_utils.cpp
         processing/component_manager.cpp
         processing/expression_node.cpp
