+++ cpp/arcticdb/processing/clause.cpp
@@ -12,7 +12,6 @@
 #include <arcticdb/processing/processing_unit.hpp>
 #include <arcticdb/column_store/segment_reslicer.hpp>
 #include <arcticdb/column_store/string_pool.hpp>
-#include <arcticdb/util/offset_string.hpp>
 #include <arcticdb/stream/merge.hpp>
 
 #include <arcticdb/processing/clause.hpp>
@@ -30,328 +29,10 @@
 #include <arcticdb/pipeline/slicing.hpp>
 #include <arcticdb/storage/store.hpp>
 #include <arcticdb/stream/stream_sink.hpp>
-#include <arcticdb/version/schema_checks.hpp>
 #include <arcticdb/pipeline/frame_utils.hpp>
 #include <arcticdb/util/collection_utils.hpp>
 
-#include <boost/regex.hpp>
-
 #include <ranges>
-
-namespace {
-using namespace arcticdb;
-/// Remove all row slice entities and ranges and keys whose indexes do not appear in row_slices_to_keep
-/// @param offsets Must be structured by row slice with ranges and keys
-/// @param ranges_and_keys Must be structured by row slice with offsets
-void filter_selected_ranges_and_keys_and_reindex_entities(
-        const std::span<const size_t> row_slices_to_keep, std::vector<std::vector<size_t>>& offsets,
-        std::vector<RangesAndKey>& ranges_and_keys
-) {
-    ARCTICDB_DEBUG_CHECK(
-            ErrorCode::E_ASSERTION_FAILURE,
-            std::ranges::adjacent_find(row_slices_to_keep, std::ranges::greater_equal{}) == row_slices_to_keep.end(),
-            "Elements of rows slices to keep must be sorted and unique"
-    );
-    std::vector<std::vector<size_t>> new_offsets;
-    new_offsets.reserve(row_slices_to_keep.size());
-    for (const size_t row_slice_to_keep : row_slices_to_keep) {
-        new_offsets.emplace_back(std::move(offsets[row_slice_to_keep]));
-    }
-    offsets = std::move(new_offsets);
-    size_t new_entity_id = 0;
-    std::vector<RangesAndKey> new_ranges_and_keys;
-    new_ranges_and_keys.reserve(ranges_and_keys.size());
-    for (std::span<size_t> row_slice : offsets) {
-        for (size_t& entity_id : row_slice) {
-            new_ranges_and_keys.emplace_back(std::move(ranges_and_keys[entity_id]));
-            entity_id = new_entity_id++;
-        }
-    }
-    ranges_and_keys = std::move(new_ranges_and_keys);
-}
-
-template<typename TDT>
-requires util::instantiation_of<TDT, TypeDescriptorTag>
-position_t write_py_string_to_pool_or_throw(
-        PyObject* const py_string_object, const size_t row_in_segment, const RowRange& row_range,
-        std::optional<ScopedGILLock>& gil_lock, StringPool& new_string_pool, const std::string_view column_name
-) {
-    return util::variant_match(
-            add_py_string_to_pool<TDT::data_type()>(py_string_object, gil_lock, new_string_pool),
-            [](position_t offset) { return offset; },
-            [&](convert::StringEncodingError&& err) -> position_t {
-                err.row_index_in_slice_ = row_in_segment;
-                err.raise(column_name, row_range.first);
-            }
-    );
-}
-
-template<typename TDT>
-requires util::instantiation_of<TDT, TypeDescriptorTag>
-void rebuild_sequence_column_in_new_pool(
-        Column& target_column, const StringPool& old_string_pool, StringPool& new_string_pool,
-        const util::BitSet* rows_to_add_in_new_pool = nullptr
-) {
-    if (rows_to_add_in_new_pool) {
-        ColumnData target_column_data = target_column.data();
-        auto accessor = random_accessor<TDT>(&target_column_data);
-        for (auto row = rows_to_add_in_new_pool->first(); row != rows_to_add_in_new_pool->end(); ++row) {
-            if (is_a_string(accessor[*row])) {
-                const std::string_view string_value = old_string_pool.get_const_view(accessor[*row]);
-                const OffsetString& new_offset = new_string_pool.get(string_value);
-                accessor[*row] = new_offset.offset();
-            }
-        }
-    } else {
-        arcticdb::for_each<TDT>(target_column, [&](auto& value) {
-            if (is_a_string(value)) {
-                const std::string_view string_value = old_string_pool.get_const_view(value);
-                const OffsetString& new_offset = new_string_pool.get(string_value);
-                value = new_offset.offset();
-            }
-        });
-    }
-}
-
-/// When merge update is performed on a string column the string pool is rebuild from scratch, thus we must iterate over
-/// all rows in the column. When the column is a part of a timeseries all target rows stored in rows_to_update will be
-/// ordered (i.e. sorted(flatten(rows_to_update)) = true). This means we can iterate over the target column row and
-/// at the same time forward iterate on the source rows. It's a nested loop but the complexity is O(n + m + k)
-/// where n is the total number of rows in the target, m are the rows in the source, and k are the total matches between
-/// source and target (i.e. sum(rows_to_update[i] for i in range(len(rows_to_update))))
-template<typename TDT>
-requires util::instantiation_of<TDT, TypeDescriptorTag>
-bool merge_update_string_column_timeseries(
-        Column& target_column, const std::span<const std::vector<size_t>> rows_to_update, const Field& target_field,
-        const RowRange& row_range, const StringPool& target_string_pool,
-        const std::span<PyObject* const> source_column_view, StringPool& new_string_pool
-) {
-    bool is_column_changed = false;
-    if constexpr (is_fixed_string_type(TDT::data_type())) {
-        user_input::raise<ErrorCode::E_INVALID_USER_ARGUMENT>(
-                "Fixed string sequences are not supported for merge update"
-        );
-    } else if constexpr (is_dynamic_string_type(TDT::data_type())) {
-        // GIL will be acquired if there is a string that is not pure ASCII/UTF-8
-        // In this case a PyObject will be allocated by convert::py_unicode_to_buffer
-        // If such a string is encountered in a column, then the GIL will be held until that whole column has
-        // been processed, on the assumption that if a column has one such string it will probably have many.
-        std::optional<ScopedGILLock> gil_lock;
-        // TODO: Create new column in case of insertion
-        auto source_row_it = rows_to_update.begin();
-        auto next_target_row_to_update = source_row_it->begin();
-        // TODO: Handle sparse target columns Monday 10756321294
-        arcticdb::for_each_enumerated<TDT>(target_column, [&](auto row) {
-            source_row_it = std::find_if(
-                    source_row_it,
-                    rows_to_update.end(),
-                    [&, new_row = false](const std::vector<size_t>& vec) mutable {
-                        // TODO: Handle inserts. Empty vec means insert.
-                        next_target_row_to_update = std::find_if(
-                                new_row ? vec.begin() : next_target_row_to_update,
-                                vec.end(),
-                                [&](const int64_t target_row_to_update) { return target_row_to_update >= row.idx(); }
-                        );
-                        if (next_target_row_to_update == vec.end()) {
-                            new_row = true;
-                        }
-                        return next_target_row_to_update != vec.end();
-                    }
-            );
-            const bool current_target_row_is_matched = source_row_it != rows_to_update.end() &&
-                                                       next_target_row_to_update != source_row_it->end() &&
-                                                       static_cast<int64_t>(*next_target_row_to_update) == row.idx();
-            if (current_target_row_is_matched) {
-                is_column_changed = true;
-                const auto py_string_object = source_column_view[source_row_it - rows_to_update.begin()];
-                row.value() = write_py_string_to_pool_or_throw<TDT>(
-                        py_string_object, row.idx(), row_range, gil_lock, new_string_pool, target_field.name()
-                );
-            } else {
-                if (is_a_string(row.value())) {
-                    const std::string_view string_in_target = target_string_pool.get_const_view(row.value());
-                    const OffsetString offset_in_new_pool = new_string_pool.get(string_in_target);
-                    row.value() = offset_in_new_pool.offset();
-                }
-            }
-        });
-    }
-    return is_column_changed;
-}
-
-/// When merge update is performed on a string column the string pool is rebuild from scratch, thus we must iterate over
-/// all rows in the column. When the column is *not* part of a timeseries the rows in rows_to_update are in random
-/// order. To avoid having quadratic complexity (i.e. O(n * k) where n are the rows in the target and k are all matches
-/// i.e. sum(rows_to_update[i] for i in range(len(rows_to_update)))) we invert the loops and iterate over all
-/// rows_to_update perform the update and then add additional loop over all rows in target to rebuild the string pool.
-/// Asymptotically this will yield the same complexity as the timeseries case but practically is a bit slower.
-template<typename TDT>
-requires util::instantiation_of<TDT, TypeDescriptorTag>
-bool merge_update_string_column_rowrange(
-        Column& target_column, const std::span<const std::vector<size_t>> rows_to_update, const Field& target_field,
-        const RowRange& row_range, const StringPool& target_string_pool,
-        const std::span<PyObject* const> source_column_view, StringPool& new_string_pool
-) {
-    bool is_column_changed = false;
-    if constexpr (is_fixed_string_type(TDT::data_type())) {
-        user_input::raise<ErrorCode::E_INVALID_USER_ARGUMENT>(
-                "Fixed string sequences are not supported for merge update"
-        );
-    } else if constexpr (is_dynamic_string_type(TDT::data_type())) {
-        util::BitSet target_rows_not_matched_by_source(target_column.row_count());
-        // GIL will be acquired if there is a string that is not pure ASCII/UTF-8
-        // In this case a PyObject will be allocated by convert::py_unicode_to_buffer
-        // If such a string is encountered in a column, then the GIL will be held until that whole column has
-        // been processed, on the assumption that if a column has one such string it will probably have many.
-        std::optional<ScopedGILLock> gil_lock;
-        ColumnData target_column_data = target_column.data();
-        auto target_column_accessor = random_accessor<TDT>(&target_column_data);
-        for (size_t source_row = 0; source_row < rows_to_update.size(); ++source_row) {
-            for (size_t target_row : rows_to_update[source_row]) {
-                const auto py_string_object = source_column_view[source_row];
-                target_column_accessor[target_row] = write_py_string_to_pool_or_throw<TDT>(
-                        py_string_object, target_row, row_range, gil_lock, new_string_pool, target_field.name()
-                );
-                is_column_changed = true;
-                target_rows_not_matched_by_source.set(target_row);
-            }
-        }
-        target_rows_not_matched_by_source.flip();
-        rebuild_sequence_column_in_new_pool<TDT>(
-                target_column, target_string_pool, new_string_pool, &target_rows_not_matched_by_source
-        );
-    }
-    return is_column_changed;
-}
-
-struct MergeUpdateStringColumnFlags {
-    /// The column is part of a timeseries dataframe
-    bool is_timeseries = false;
-    /// Only move the data to a new string pool.
-    bool data_not_changed = false;
-};
-
-template<typename TDT>
-requires util::instantiation_of<TDT, TypeDescriptorTag>
-bool merge_update_string_column(
-        Column& target_column, const MergeUpdateStringColumnFlags& flags,
-        const std::span<const std::vector<size_t>> rows_to_update, const Field& target_field, const RowRange& row_range,
-        const StringPool& target_string_pool, const std::span<PyObject* const> source_column_view,
-        StringPool& new_string_pool
-) {
-    if (flags.data_not_changed) {
-        rebuild_sequence_column_in_new_pool<TDT>(target_column, target_string_pool, new_string_pool);
-        return false;
-    } else {
-        if (flags.is_timeseries) {
-            return merge_update_string_column_timeseries<TDT>(
-                    target_column,
-                    rows_to_update,
-                    target_field,
-                    row_range,
-                    target_string_pool,
-                    source_column_view,
-                    new_string_pool
-            );
-        } else {
-            return merge_update_string_column_rowrange<TDT>(
-                    target_column,
-                    rows_to_update,
-                    target_field,
-                    row_range,
-                    target_string_pool,
-                    source_column_view,
-                    new_string_pool
-            );
-        }
-    }
-}
-
-struct NaNAwareFloatComparator {
-    template<std::floating_point T>
-    bool operator()(const T a, const T b) const {
-        return a == b || (std::isnan(a) && std::isnan(b));
-    }
-};
-
-struct NaNAwareFloatHasher {
-    using is_avalanching = void;
-    template<std::floating_point T>
-    uint64_t operator()(const T a) const {
-        if (std::isnan(a)) {
-            // IEEE allows multiple different bit representations of NaN. std::isnan is required to return true for all
-            // different bit bit representations of NaN, std::quient_NaN is an implementation defined constant.
-            return ankerl::unordered_dense::hash<T>()(std::numeric_limits<T>::quiet_NaN());
-        } else {
-            return std::hash<T>()(a);
-        }
-    }
-};
-
-template<
-        typename SourceTDT, typename TargetTDT, typename SourceValueRawType = TargetTDT::DataTypeTag::raw_type,
-        typename TargetValueRawType = TargetTDT::DataTypeTag::raw_type>
-requires std::same_as<std::decay_t<SourceTDT>, std::decay_t<TargetTDT>>
-std::variant<bool, convert::StringEncodingError> are_merge_values_matching(
-        const SourceValueRawType& source_value, const TargetValueRawType& target_value,
-        const StringPool& target_string_pool, std::optional<ScopedGILLock>& scoped_gil_lock
-) {
-    if constexpr (is_sequence_type(SourceTDT::data_type())) {
-        const bool is_source_null = is_py_none(source_value) || is_py_nan(source_value);
-        const bool is_target_null = !is_a_string(target_value);
-        if (is_source_null ^ is_target_null) {
-            return false;
-        } else if (is_source_null && is_target_null) {
-            return true;
-        } else {
-            return util::variant_match(
-                    create_py_object_wrapper_or_error<TargetTDT::data_type()>(source_value, scoped_gil_lock),
-                    [](convert::StringEncodingError&& err) -> std::variant<bool, convert::StringEncodingError> {
-                        return err;
-                    },
-                    [&](convert::PyStringWrapper&& wrapper) -> std::variant<bool, convert::StringEncodingError> {
-                        return target_string_pool.get_const_view(target_value) ==
-                               std::string_view(wrapper.buffer_, wrapper.length_);
-                    }
-            );
-        }
-    } else if constexpr (is_floating_point_type(SourceTDT::data_type())) {
-        constexpr static NaNAwareFloatComparator comparator;
-        return comparator(source_value, target_value);
-    } else {
-        return source_value == target_value;
-    }
-}
-
-template<typename TDT>
-requires util::instantiation_of<TDT, TypeDescriptorTag>
-auto map_column_values_to_rows(const ColumnWithStrings& column) {
-    constexpr static bool is_target_sequence_type = is_sequence_type(TDT::data_type());
-    constexpr static bool is_target_floating_point_type = is_floating_point_type(TDT::data_type());
-    using TargetRawType = typename TDT::DataTypeTag::raw_type;
-    using TargetValueType = std::conditional_t<is_target_sequence_type, std::optional<std::string_view>, TargetRawType>;
-    using Hasher = std::conditional_t<
-            is_target_floating_point_type,
-            NaNAwareFloatHasher,
-            ankerl::unordered_dense::hash<TargetValueType>>;
-    using Comparator =
-            std::conditional_t<is_target_floating_point_type, NaNAwareFloatComparator, std::equal_to<TargetValueType>>;
-    ankerl::unordered_dense::map<TargetValueType, std::vector<size_t>, Hasher, Comparator> target_values;
-    arcticdb::for_each_enumerated<TDT>(*column.column_, [&](auto row) {
-        if constexpr (is_target_sequence_type) {
-            if (is_a_string(row.value())) {
-                target_values[column.string_at_offset(row.value())].push_back(row.idx());
-            } else {
-                target_values[std::nullopt].push_back(row.idx());
-            }
-        } else {
-            target_values[row.value()].push_back(row.idx());
-        }
-    });
-    return target_values;
-}
-
-} // namespace
 
 namespace arcticdb {
 
@@ -416,41 +97,6 @@ struct SegmentWrapper {
 
     const StreamId& id() const { return id_; }
 };
-
-static auto first_missing_column(OutputSchema& output_schema, const std::unordered_set<std::string>& required_columns) {
-    const auto& column_types = output_schema.column_types();
-    for (auto input_column_it = required_columns.begin(); input_column_it != required_columns.end();
-         ++input_column_it) {
-        if (!column_types.contains(*input_column_it) &&
-            !column_types.contains(stream::mangled_name(*input_column_it))) {
-            return input_column_it;
-        }
-    }
-    return required_columns.end();
-}
-
-void check_column_presence(
-        OutputSchema& output_schema, const std::unordered_set<std::string>& required_columns,
-        std::string_view clause_name
-) {
-    const auto first_missing = first_missing_column(output_schema, required_columns);
-    schema::check<ErrorCode::E_COLUMN_DOESNT_EXIST>(
-            first_missing == required_columns.end(),
-            "{}Clause requires column '{}' to exist in input data",
-            clause_name,
-            first_missing == required_columns.end() ? "" : *first_missing
-    );
-}
-
-void check_is_timeseries(const StreamDescriptor& stream_descriptor, std::string_view clause_name) {
-    schema::check<ErrorCode::E_UNSUPPORTED_INDEX_TYPE>(
-            stream_descriptor.index().type() == IndexDescriptor::Type::TIMESTAMP &&
-                    stream_descriptor.index().field_count() >= 1 &&
-                    stream_descriptor.field(0).type() == make_scalar_type(DataType::NANOSECONDS_UTC64),
-            "{}Clause can only be applied to timeseries",
-            clause_name
-    );
-}
 
 OutputSchema modify_schema(OutputSchema&& schema, const std::vector<std::shared_ptr<Clause>>& clauses) {
     for (const auto& clause : clauses) {
@@ -929,490 +575,6 @@ OutputSchema AggregationClause::modify_schema(OutputSchema&& output_schema) cons
 }
 
 [[nodiscard]] std::string AggregationClause::to_string() const { return str_; }
-
-template<ResampleBoundary closed_boundary>
-ResampleClause<closed_boundary>::ResampleClause(
-        std::string rule, ResampleBoundary label_boundary, BucketGeneratorT&& generate_bucket_boundaries,
-        timestamp offset, ResampleOrigin origin
-) :
-    rule_(std::move(rule)),
-    label_boundary_(label_boundary),
-    generate_bucket_boundaries_(std::move(generate_bucket_boundaries)),
-    offset_(offset),
-    origin_(std::move(origin)) {
-    clause_info_.input_structure_ = ProcessingStructure::TIME_BUCKETED;
-    clause_info_.can_combine_with_column_selection_ = false;
-    clause_info_.index_ = KeepCurrentTopLevelIndex();
-}
-
-template<ResampleBoundary closed_boundary>
-const ClauseInfo& ResampleClause<closed_boundary>::clause_info() const {
-    return clause_info_;
-}
-
-template<ResampleBoundary closed_boundary>
-void ResampleClause<closed_boundary>::set_component_manager(std::shared_ptr<ComponentManager> component_manager) {
-    component_manager_ = std::move(component_manager);
-}
-
-template<ResampleBoundary closed_boundary>
-OutputSchema ResampleClause<closed_boundary>::modify_schema(OutputSchema&& output_schema) const {
-    check_is_timeseries(output_schema.stream_descriptor(), "Resample");
-    output_schema.clear_default_values();
-    check_column_presence(output_schema, *clause_info_.input_columns_, "Resample");
-    const auto& input_stream_desc = output_schema.stream_descriptor();
-    StreamDescriptor stream_desc(input_stream_desc.id());
-    stream_desc.add_field(input_stream_desc.field(0));
-    stream_desc.set_index(IndexDescriptorImpl(IndexDescriptor::Type::TIMESTAMP, 1));
-
-    for (const auto& agg : aggregators_) {
-        const auto& input_column_name = agg.get_input_column_name().value;
-        const auto& output_column_name = agg.get_output_column_name().value;
-        const auto& input_column_type = output_schema.column_types()[input_column_name];
-        agg.check_aggregator_supported_with_data_type(input_column_type);
-        auto output_column_type = agg.generate_output_data_type(input_column_type);
-        stream_desc.add_scalar_field(output_column_type, output_column_name);
-        const std::optional<Value>& default_value = agg.get_default_value(input_column_type);
-        if (default_value) {
-            output_schema.set_default_value_for_column(output_column_name, *default_value);
-        }
-    }
-    output_schema.set_stream_descriptor(std::move(stream_desc));
-
-    if (output_schema.norm_metadata_.df().common().has_multi_index()) {
-        const auto& multi_index = output_schema.norm_metadata_.mutable_df()->mutable_common()->multi_index();
-        auto name = multi_index.name();
-        auto tz = multi_index.tz();
-        bool fake_name{false};
-        for (auto pos : multi_index.fake_field_pos()) {
-            if (pos == 0) {
-                fake_name = true;
-                break;
-            }
-        }
-        auto mutable_index = output_schema.norm_metadata_.mutable_df()->mutable_common()->mutable_index();
-        mutable_index->set_tz(tz);
-        mutable_index->set_is_physically_stored(true);
-        mutable_index->set_name(name);
-        mutable_index->set_fake_name(fake_name);
-    }
-    return output_schema;
-}
-
-template<ResampleBoundary closed_boundary>
-std::string ResampleClause<closed_boundary>::rule() const {
-    return rule_;
-}
-
-template<ResampleBoundary closed_boundary>
-void ResampleClause<closed_boundary>::set_date_range(timestamp date_range_start, timestamp date_range_end) {
-    // Start and end need to read the first and last segments of the date range. At the moment buckets are set up
-    // before reading and processing the data.
-    constexpr static std::array unsupported_origin{"start", "end", "start_day", "end_day"};
-    user_input::check<ErrorCode::E_INVALID_USER_ARGUMENT>(
-            util::variant_match(
-                    origin_,
-                    [&](const std::string& origin) {
-                        return ranges::none_of(unsupported_origin, [&](std::string_view el) { return el == origin; });
-                    },
-                    [](const auto&) { return true; }
-            ),
-            "Resampling origins {} are not supported in conjunction with date range",
-            unsupported_origin
-    );
-    date_range_.emplace(date_range_start, date_range_end);
-}
-
-template<ResampleBoundary closed_boundary>
-void ResampleClause<closed_boundary>::set_aggregations(const std::vector<NamedAggregator>& named_aggregators) {
-    clause_info_.input_columns_ = std::make_optional<std::unordered_set<std::string>>();
-    str_ = fmt::format("RESAMPLE({}) | AGGREGATE {{", rule());
-    for (const auto& named_aggregator : named_aggregators) {
-        str_.append(fmt::format(
-                "{}: ({}, {}), ",
-                named_aggregator.output_column_name_,
-                named_aggregator.input_column_name_,
-                named_aggregator.aggregation_operator_
-        ));
-        clause_info_.input_columns_->insert(named_aggregator.input_column_name_);
-        auto typed_input_column_name = ColumnName(named_aggregator.input_column_name_);
-        auto typed_output_column_name = ColumnName(named_aggregator.output_column_name_);
-        if (named_aggregator.aggregation_operator_ == "sum") {
-            aggregators_.emplace_back(SortedAggregator<AggregationOperator::SUM, closed_boundary>(
-                    typed_input_column_name, typed_output_column_name
-            ));
-        } else if (named_aggregator.aggregation_operator_ == "mean") {
-            aggregators_.emplace_back(SortedAggregator<AggregationOperator::MEAN, closed_boundary>(
-                    typed_input_column_name, typed_output_column_name
-            ));
-        } else if (named_aggregator.aggregation_operator_ == "min") {
-            aggregators_.emplace_back(SortedAggregator<AggregationOperator::MIN, closed_boundary>(
-                    typed_input_column_name, typed_output_column_name
-            ));
-        } else if (named_aggregator.aggregation_operator_ == "max") {
-            aggregators_.emplace_back(SortedAggregator<AggregationOperator::MAX, closed_boundary>(
-                    typed_input_column_name, typed_output_column_name
-            ));
-        } else if (named_aggregator.aggregation_operator_ == "first") {
-            aggregators_.emplace_back(SortedAggregator<AggregationOperator::FIRST, closed_boundary>(
-                    typed_input_column_name, typed_output_column_name
-            ));
-        } else if (named_aggregator.aggregation_operator_ == "last") {
-            aggregators_.emplace_back(SortedAggregator<AggregationOperator::LAST, closed_boundary>(
-                    typed_input_column_name, typed_output_column_name
-            ));
-        } else if (named_aggregator.aggregation_operator_ == "count") {
-            aggregators_.emplace_back(SortedAggregator<AggregationOperator::COUNT, closed_boundary>(
-                    typed_input_column_name, typed_output_column_name
-            ));
-        } else {
-            user_input::raise<ErrorCode::E_INVALID_USER_ARGUMENT>(
-                    "Unknown aggregation operator provided to resample: {}", named_aggregator.aggregation_operator_
-            );
-        }
-    }
-    str_.append("}");
-}
-
-template<ResampleBoundary closed_boundary>
-void ResampleClause<closed_boundary>::set_processing_config(const ProcessingConfig& processing_config) {
-    processing_config_ = processing_config;
-}
-
-template<ResampleBoundary closed_boundary>
-std::vector<std::vector<size_t>> ResampleClause<closed_boundary>::structure_for_processing(
-        std::vector<RangesAndKey>& ranges_and_keys
-) {
-    ARCTICDB_RUNTIME_DEBUG(log::memory(), "ResampleClause: structure for processing 1");
-    if (ranges_and_keys.empty()) {
-        return {};
-    }
-    user_input::check<ErrorCode::E_INVALID_USER_ARGUMENT>(
-            processing_config_.index_type_ == IndexDescriptor::Type::TIMESTAMP,
-            "Cannot resample non-timestamp indexed data"
-    );
-
-    // Iterate over ranges_and_keys and create a pair with first element equal to the smallest start time and second
-    // element equal to the largest end time.
-    const TimestampRange index_range = std::accumulate(
-            std::next(ranges_and_keys.begin()),
-            ranges_and_keys.end(),
-            TimestampRange{ranges_and_keys.begin()->start_time(), ranges_and_keys.begin()->end_time()},
-            [](const TimestampRange& rng, const RangesAndKey& el) {
-                return TimestampRange{std::min(rng.first, el.start_time()), std::max(rng.second, el.end_time())};
-            }
-    );
-
-    if (date_range_.has_value()) {
-        date_range_->first = std::max(date_range_->first, index_range.first);
-        date_range_->second = std::min(date_range_->second, index_range.second);
-    } else {
-        date_range_ = index_range;
-    }
-
-    bucket_boundaries_ = generate_bucket_boundaries_(
-            date_range_->first, date_range_->second, rule_, closed_boundary, offset_, origin_
-    );
-    if (bucket_boundaries_.size() < 2) {
-        ranges_and_keys.clear();
-        return {};
-    }
-    ARCTICDB_DEBUG_CHECK(
-            ErrorCode::E_ASSERTION_FAILURE,
-            ranges::is_sorted(bucket_boundaries_),
-            "Resampling expects provided bucket boundaries to be strictly monotonically increasing"
-    );
-    return structure_by_time_bucket<closed_boundary>(ranges_and_keys, bucket_boundaries_);
-}
-
-template<ResampleBoundary closed_boundary>
-std::vector<std::vector<EntityId>> ResampleClause<closed_boundary>::structure_for_processing(
-        std::vector<std::vector<EntityId>>&& entity_ids_vec
-) {
-    auto entity_ids = util::flatten_vectors(std::move(entity_ids_vec));
-    if (entity_ids.empty()) {
-        return {};
-    }
-    ARCTICDB_RUNTIME_DEBUG(log::memory(), "ResampleClause: structure for processing 2");
-    auto [segments, row_ranges, col_ranges] = component_manager_->get_entities<
-            std::shared_ptr<SegmentInMemory>,
-            std::shared_ptr<RowRange>,
-            std::shared_ptr<ColRange>>(entity_ids);
-    std::vector<RangesAndEntity> ranges_and_entities;
-    ranges_and_entities.reserve(entity_ids.size());
-    timestamp min_start_ts{std::numeric_limits<timestamp>::max()};
-    timestamp max_end_ts{std::numeric_limits<timestamp>::min()};
-    for (size_t idx = 0; idx < entity_ids.size(); ++idx) {
-        auto start_ts = std::get<timestamp>(stream::TimeseriesIndex::start_value_for_segment(*segments[idx]));
-        auto end_ts = std::get<timestamp>(stream::TimeseriesIndex::end_value_for_segment(*segments[idx]));
-        min_start_ts = std::min(min_start_ts, start_ts);
-        max_end_ts = std::max(max_end_ts, end_ts);
-        ranges_and_entities.emplace_back(
-                entity_ids[idx], row_ranges[idx], col_ranges[idx], std::make_optional<TimestampRange>(start_ts, end_ts)
-        );
-    }
-
-    date_range_ = std::make_optional<TimestampRange>(min_start_ts, max_end_ts);
-    bucket_boundaries_ = generate_bucket_boundaries_(
-            date_range_->first, date_range_->second, rule_, closed_boundary, offset_, origin_
-    );
-    if (bucket_boundaries_.size() < 2) {
-        return {};
-    }
-    ARCTICDB_DEBUG_CHECK(
-            ErrorCode::E_ASSERTION_FAILURE,
-            ranges::is_sorted(bucket_boundaries_),
-            "Resampling expects provided bucket boundaries to be strictly monotonically increasing"
-    );
-
-    auto new_structure_offsets = structure_by_time_bucket<closed_boundary>(ranges_and_entities, bucket_boundaries_);
-
-    std::vector<EntityFetchCount> expected_fetch_counts(ranges_and_entities.size(), 0);
-    for (const auto& list : new_structure_offsets) {
-        for (auto idx : list) {
-            internal::check<ErrorCode::E_ASSERTION_FAILURE>(
-                    idx < expected_fetch_counts.size(),
-                    "Index {} in new_structure_offsets out of bounds >{}",
-                    idx,
-                    expected_fetch_counts.size() - 1
-            );
-            expected_fetch_counts[idx]++;
-        }
-    }
-    internal::check<ErrorCode::E_ASSERTION_FAILURE>(
-            ranges::all_of(
-                    expected_fetch_counts,
-                    [](EntityFetchCount fetch_count) { return fetch_count == 1 || fetch_count == 2; }
-            ),
-            "ResampleClause::structure_for_processing: invalid expected entity fetch count (should be 1 or 2)"
-    );
-    std::vector<EntityId> entities_to_be_fetched_twice;
-    for (auto&& [idx, ranges_and_entity] : folly::enumerate(ranges_and_entities)) {
-        if (expected_fetch_counts[idx] == 2) {
-            entities_to_be_fetched_twice.emplace_back(ranges_and_entity.id_);
-        }
-    }
-    component_manager_->replace_entities<EntityFetchCount>(entities_to_be_fetched_twice, EntityFetchCount(2));
-    return offsets_to_entity_ids(new_structure_offsets, ranges_and_entities);
-}
-
-template<ResampleBoundary closed_boundary>
-std::vector<EntityId> ResampleClause<closed_boundary>::process(std::vector<EntityId>&& entity_ids) const {
-    ARCTICDB_SAMPLE(ResampleClause, 0)
-    if (entity_ids.empty()) {
-        return {};
-    }
-    auto proc = gather_entities<
-            std::shared_ptr<SegmentInMemory>,
-            std::shared_ptr<RowRange>,
-            std::shared_ptr<ColRange>,
-            EntityFetchCount>(*component_manager_, std::move(entity_ids));
-    ARCTICDB_RUNTIME_DEBUG(log::memory(), "ResampleClause: processing entities {}", entity_ids);
-    auto row_slices = split_by_row_slice(std::move(proc));
-    // If the entity fetch counts for the entities in the first row slice are 2, the first bucket overlapping this
-    // row slice is being computed by the call to process dealing with the row slices above these. Otherwise, this
-    // call should do it
-    const auto& front_slice = row_slices.front();
-    const bool responsible_for_first_overlapping_bucket = front_slice.entity_fetch_count_->at(0) == 1;
-    // Find the iterators into bucket_boundaries_ of the start of the first and the end of the last bucket this call
-    // to process is responsible for calculating All segments in a given row slice contain the same index column, so
-    // just grab info from the first one
-    const auto& index_column_name = front_slice.segments_->at(0)->field(0).name();
-    const auto& first_row_slice_index_col = front_slice.segments_->at(0)->column(0);
-    // Resampling only makes sense for timestamp indexes
-    internal::check<ErrorCode::E_ASSERTION_FAILURE>(
-            is_time_type(first_row_slice_index_col.type().data_type()),
-            "Cannot resample data with index column of non-timestamp type"
-    );
-    const auto first_ts = first_row_slice_index_col.template scalar_at<timestamp>(0).value();
-    // If there is only one row slice, then the last index value of interest is just the last index value for this
-    // row slice. If there is more than one, then the first index value from the second row slice must be used to
-    // calculate the buckets of interest, due to an old bug in update. See
-    // test_compatibility.py::test_compat_resample_updated_data for details
-    const auto last_ts =
-            row_slices.size() == 1
-                    ? first_row_slice_index_col.template scalar_at<timestamp>(first_row_slice_index_col.row_count() - 1)
-                              .value()
-                    : row_slices.back().segments_->at(0)->column(0).template scalar_at<timestamp>(0).value();
-    auto bucket_boundaries = generate_bucket_boundaries(first_ts, last_ts, responsible_for_first_overlapping_bucket);
-    if (bucket_boundaries.size() < 2) {
-        return {};
-    }
-    std::vector<std::shared_ptr<Column>> input_index_columns;
-    input_index_columns.reserve(row_slices.size());
-    for (const auto& row_slice : row_slices) {
-        input_index_columns.emplace_back(row_slice.segments_->at(0)->column_ptr(0));
-    }
-    const auto output_index_column = generate_output_index_column(input_index_columns, bucket_boundaries);
-    // Bucket boundaries can be wider than the date range specified by the user, narrow the first and last buckets
-    // here if necessary
-    bucket_boundaries.front(
-    ) = std::max(bucket_boundaries.front(), date_range_->first - (closed_boundary == ResampleBoundary::RIGHT ? 1 : 0));
-    bucket_boundaries.back(
-    ) = std::min(bucket_boundaries.back(), date_range_->second + (closed_boundary == ResampleBoundary::LEFT ? 1 : 0));
-    SegmentInMemory seg;
-    RowRange output_row_range(
-            row_slices.front().row_ranges_->at(0)->start(),
-            row_slices.front().row_ranges_->at(0)->start() + output_index_column->row_count()
-    );
-    ColRange output_col_range(1, aggregators_.size() + 1);
-    seg.add_column(scalar_field(DataType::NANOSECONDS_UTC64, index_column_name), output_index_column);
-    seg.descriptor().set_index(IndexDescriptorImpl(IndexDescriptor::Type::TIMESTAMP, 1));
-    auto& string_pool = seg.string_pool();
-
-    ARCTICDB_DEBUG_THROW(5)
-    for (const auto& aggregator : aggregators_) {
-        std::vector<std::optional<ColumnWithStrings>> input_agg_columns;
-        input_agg_columns.reserve(row_slices.size());
-        for (auto& row_slice : row_slices) {
-            auto variant_data = row_slice.get(aggregator.get_input_column_name());
-            util::variant_match(
-                    variant_data,
-                    [&input_agg_columns](const ColumnWithStrings& column_with_strings) {
-                        input_agg_columns.emplace_back(column_with_strings);
-                    },
-                    [&input_agg_columns](const EmptyResult&) {
-                        // Dynamic schema, missing column from this row-slice
-                        // Not currently supported, but will be, hence the argument to aggregate being a vector of
-                        // optionals
-                        input_agg_columns.emplace_back();
-                    },
-                    [](const auto&) {
-                        internal::raise<ErrorCode::E_ASSERTION_FAILURE>(
-                                "Unexpected return type from ProcessingUnit::get, expected column-like"
-                        );
-                    }
-            );
-        }
-        std::optional<Column> aggregated = aggregator.aggregate(
-                input_index_columns,
-                input_agg_columns,
-                bucket_boundaries,
-                *output_index_column,
-                string_pool,
-                label_boundary_
-        );
-        if (aggregated) {
-            seg.add_column(
-                    scalar_field(aggregated->type().data_type(), aggregator.get_output_column_name().value),
-                    std::make_shared<Column>(std::move(aggregated).value())
-            );
-        }
-    }
-    seg.set_row_data(output_index_column->row_count() - 1);
-    return push_entities(
-            *component_manager_,
-            ProcessingUnit(std::move(seg), std::move(output_row_range), std::move(output_col_range))
-    );
-}
-
-template<ResampleBoundary closed_boundary>
-[[nodiscard]] std::string ResampleClause<closed_boundary>::to_string() const {
-    return str_;
-}
-
-template<ResampleBoundary closed_boundary>
-std::vector<timestamp> ResampleClause<closed_boundary>::generate_bucket_boundaries(
-        timestamp first_ts, timestamp last_ts, bool responsible_for_first_overlapping_bucket
-) const {
-    auto first_it = std::lower_bound(
-            bucket_boundaries_.begin(),
-            bucket_boundaries_.end(),
-            first_ts,
-            [](timestamp boundary, timestamp first_ts) {
-                if constexpr (closed_boundary == ResampleBoundary::LEFT) {
-                    return boundary <= first_ts;
-                } else {
-                    // closed_boundary == ResampleBoundary::RIGHT
-                    return boundary < first_ts;
-                }
-            }
-    );
-    if (responsible_for_first_overlapping_bucket && first_it != bucket_boundaries_.begin()) {
-        --first_it;
-    }
-    auto last_it =
-            std::upper_bound(first_it, bucket_boundaries_.end(), last_ts, [](timestamp last_ts, timestamp boundary) {
-                if constexpr (closed_boundary == ResampleBoundary::LEFT) {
-                    return last_ts < boundary;
-                } else {
-                    // closed_boundary == ResampleBoundary::RIGHT
-                    return last_ts <= boundary;
-                }
-            });
-    if (last_it != bucket_boundaries_.end()) {
-        ++last_it;
-    }
-    std::vector<timestamp> bucket_boundaries(first_it, last_it);
-    // There used to be a check here that there was at least one bucket to process. However, this is not always the
-    // case for data written by old versions of Arctic using update. See
-    // test_compatibility.py::test_compat_resample_updated_data for more explanation
-    return bucket_boundaries;
-}
-
-template<ResampleBoundary closed_boundary>
-std::shared_ptr<Column> ResampleClause<closed_boundary>::generate_output_index_column(
-        const std::vector<std::shared_ptr<Column>>& input_index_columns, const std::vector<timestamp>& bucket_boundaries
-) const {
-    constexpr auto data_type = DataType::NANOSECONDS_UTC64;
-    using IndexTDT = ScalarTagType<DataTypeTag<data_type>>;
-
-    const auto max_index_column_bytes = (bucket_boundaries.size() - 1) * get_type_size(data_type);
-    auto output_index_column = std::make_shared<Column>(
-            TypeDescriptor(data_type, Dimension::Dim0),
-            Sparsity::NOT_PERMITTED,
-            ChunkedBuffer::presized_in_blocks(max_index_column_bytes)
-    );
-    auto output_index_column_data = output_index_column->data();
-    auto output_index_column_it = output_index_column_data.template begin<IndexTDT>();
-    size_t output_index_column_row_count{0};
-
-    auto bucket_end_it = std::next(bucket_boundaries.cbegin());
-    Bucket<closed_boundary> current_bucket{*std::prev(bucket_end_it), *bucket_end_it};
-    bool current_bucket_added_to_index{false};
-    // Only include buckets that have at least one index value in range
-    for (const auto& input_index_column : input_index_columns) {
-        auto index_column_data = input_index_column->data();
-        const auto cend = index_column_data.cend<IndexTDT>();
-        auto it = index_column_data.cbegin<IndexTDT>();
-        // In case the passed date_range does not span the whole segment we need to skip the index values
-        // which are before the date range start.
-        while (it != cend && *it < date_range_->first) {
-            ++it;
-        }
-        for (; it != cend && *it <= date_range_->second; ++it) {
-            if (ARCTICDB_LIKELY(current_bucket.contains(*it))) {
-                if (ARCTICDB_UNLIKELY(!current_bucket_added_to_index)) {
-                    *output_index_column_it++ =
-                            label_boundary_ == ResampleBoundary::LEFT ? *std::prev(bucket_end_it) : *bucket_end_it;
-                    ++output_index_column_row_count;
-                    current_bucket_added_to_index = true;
-                }
-            } else {
-                advance_boundary_past_value<closed_boundary>(bucket_boundaries, bucket_end_it, *it);
-                if (ARCTICDB_UNLIKELY(bucket_end_it == bucket_boundaries.end())) {
-                    break;
-                } else {
-                    current_bucket.set_boundaries(*std::prev(bucket_end_it), *bucket_end_it);
-                    current_bucket_added_to_index = false;
-                    if (ARCTICDB_LIKELY(current_bucket.contains(*it))) {
-                        *output_index_column_it++ =
-                                label_boundary_ == ResampleBoundary::LEFT ? *std::prev(bucket_end_it) : *bucket_end_it;
-                        ++output_index_column_row_count;
-                        current_bucket_added_to_index = true;
-                    }
-                }
-            }
-        }
-    }
-    const auto actual_index_column_bytes = output_index_column_row_count * get_type_size(data_type);
-    output_index_column->buffer().trim(actual_index_column_bytes);
-    output_index_column->set_row_data(output_index_column_row_count - 1);
-    return output_index_column;
-}
-
-template struct ResampleClause<ResampleBoundary::LEFT>;
-template struct ResampleClause<ResampleBoundary::RIGHT>;
 
 [[nodiscard]] std::vector<EntityId> RemoveColumnPartitioningClause::process(std::vector<EntityId>&& entity_ids) const {
     if (entity_ids.empty()) {
@@ -2068,582 +1230,6 @@ OutputSchema WriteClause::join_schemas(std::vector<OutputSchema>&&) const {
 }
 
 std::string WriteClause::to_string() const { return "Write"; }
-
-MergeUpdateClause::MergeUpdateClause(
-        std::vector<std::string>&& on, MergeStrategy strategy, std::shared_ptr<InputFrame> source
-) :
-
-    on_(std::move(on)),
-    strategy_(strategy),
-    source_(std::move(source)) {
-    std::erase_if(on_, [&](const std::string& column) { return !on_set_.insert(column).second; });
-    user_input::check<ErrorCode::E_INVALID_USER_ARGUMENT>(
-            strategy_.not_matched_by_target == MergeAction::DO_NOTHING, "Merge cannot perform insertion at the moment"
-    );
-}
-
-std::vector<std::vector<size_t>> MergeUpdateClause::structure_for_processing(std::vector<RangesAndKey>& ranges_and_keys
-) {
-    if (ranges_and_keys.empty()) {
-        return {};
-    }
-    if (!source_->has_index()) {
-        return structure_by_row_slice(ranges_and_keys);
-    }
-    return structure_for_processing_log(ranges_and_keys);
-}
-
-std::vector<std::vector<size_t>> MergeUpdateClause::structure_for_processing_log(
-        std::vector<RangesAndKey>& ranges_and_keys
-) {
-
-    using IndexType = ScalarTagType<DataTypeTag<DataType::NANOSECONDS_UTC64>>;
-    std::vector<std::vector<size_t>> offsets = structure_by_row_slice(ranges_and_keys);
-    std::vector<size_t> row_slices_to_keep;
-    // TODO: Arrow is not supported yet.
-    const TypedTensor<IndexType::DataTypeTag::raw_type> index_tensor(source_->opt_index_tensor().value());
-    user_input::check<ErrorCode::E_INVALID_USER_ARGUMENT>(
-            util::is_cstyle_array<IndexType::DataTypeTag::raw_type>(index_tensor),
-            "Fortran-style arrays are not supported by merge update yet. The index column has data type {} of "
-            "size {} bytes but the stride is {} bytes",
-            IndexType::data_type(),
-            sizeof(IndexType::DataTypeTag::raw_type),
-            index_tensor.strides()[0]
-    );
-    const bool source_ends_before_first_row_slice =
-            index_tensor.at(source_->num_rows - 1) < ranges_and_keys.begin()->key_.time_range().first;
-    const bool source_starts_after_the_last_row_slice =
-            index_tensor.at(0) >= ranges_and_keys.back().key_.time_range().second;
-    if (strategy_.update_only() && (source_ends_before_first_row_slice || source_starts_after_the_last_row_slice)) {
-        ranges_and_keys.clear();
-        return {};
-    }
-    std::span index(static_cast<const IndexType::DataTypeTag::raw_type*>(index_tensor.data()), source_->num_rows);
-    for (size_t row_slice_idx = 0; row_slice_idx < offsets.size(); ++row_slice_idx) {
-        // TODO: Add logic for insertion.
-        const TimestampRange& time_range = ranges_and_keys[offsets[row_slice_idx].front()].key_.time_range();
-        if (source_start_end_for_row_range_.contains(time_range)) {
-            row_slices_to_keep.push_back(row_slice_idx);
-            continue;
-        }
-        const auto source_range_start = std::ranges::lower_bound(index, time_range.first);
-        if (source_range_start == index.end()) {
-            break;
-        }
-        if (*source_range_start < time_range.second) {
-            auto source_range_end = std::upper_bound(source_range_start, index.end(), time_range.second - 1);
-            const std::pair<size_t, size_t> source_row_range = {
-                    source_range_start - index.begin(), source_range_end - index.begin()
-            };
-            source_start_end_for_row_range_.insert({time_range, source_row_range});
-            row_slices_to_keep.push_back(row_slice_idx);
-        }
-    }
-    filter_selected_ranges_and_keys_and_reindex_entities(row_slices_to_keep, offsets, ranges_and_keys);
-    return offsets;
-}
-
-std::vector<std::vector<EntityId>> MergeUpdateClause::structure_for_processing(std::vector<std::vector<EntityId>>&&) {
-    internal::raise<ErrorCode::E_ASSERTION_FAILURE>("MergeUpdate clause should be the first clause in the pipeline");
-}
-
-/// Decide which rows of should be updated and which rows from source should be inserted.
-/// 1. If there's a timestamp index  use MergeUpdateClause::filter_index_match this will produce a vector of size
-/// equal to the number of rows from the source that fall into the processed slice. Each vector will contain a
-/// vector of row-indexes in target that match the corresponding source index value.
-/// 2. For each column in MergeUpdateClause::on_ iterate over the vector of vectors produced in the previous step.
-/// Checking for match only the target rows that are in the inner vector. If there is no match for this particular
-/// column remove the target row index.
-/// This means that the ordering of the columns in MergeUpdateClause::on_ matters and it would be more efficient to
-/// start with the columns that have a lesser chance of matching.
-std::vector<EntityId> MergeUpdateClause::process(std::vector<EntityId>&& entity_ids) const {
-    if (entity_ids.empty()) {
-        return {};
-    }
-    auto proc = gather_entities<
-            std::shared_ptr<SegmentInMemory>,
-            std::shared_ptr<RowRange>,
-            std::shared_ptr<ColRange>,
-            std::shared_ptr<AtomKey>>(*component_manager_, std::move(entity_ids));
-    // TODO: Add exception handling two source rows matching the same target row. This should be done in the
-    // function
-    //  handling the "on" parameter matching multiple columns. Monday 10655943156
-    auto matched = [&]() -> std::optional<std::vector<std::vector<size_t>>> {
-        std::optional<std::vector<std::vector<size_t>>> result;
-        if (source_->has_index()) {
-            using IndexType = ScalarTagType<DataTypeTag<DataType::NANOSECONDS_UTC64>>;
-            const TypedTensor<IndexType::DataTypeTag::raw_type> index_tensor(source_->opt_index_tensor().value());
-            result.emplace(filter_index_match(
-                    proc.segments_->front()->column(0), std::span(index_tensor.data(), source_->num_rows), proc
-            ));
-        }
-        return result;
-    }();
-    // TODO: Source and target descriptor can differ for dynamic schema
-    matched = filter_on_additional_columns_match(
-            source_->desc(), source_->desc(), source_->field_tensors(), proc, std::move(matched)
-    );
-    if (source_->has_segment()) {
-        user_input::raise<ErrorCode::E_INVALID_USER_ARGUMENT>("Arrow format is not supported as input for merge update"
-        );
-    } else if (source_->has_tensors()) {
-        if (update_and_insert(source_->field_tensors(), source_->desc(), proc, *matched)) {
-            return push_entities(*component_manager_, std::move(proc));
-        }
-    } else {
-        internal::raise<ErrorCode::E_ASSERTION_FAILURE>("Input frame does not contain neither a segment nor tensors");
-    }
-    return {};
-}
-
-std::vector<std::vector<size_t>> MergeUpdateClause::initialize_rows_to_update_for_rowrange_indexed_data(
-        ProcessingUnit& proc, const StreamDescriptor& source_descriptor
-) const {
-    user_input::check<ErrorCode::E_INVALID_USER_ARGUMENT>(
-            !on_.empty(),
-            "MergeUpdate requires at least one column in the \"on\" parameter when the DataFrame is not a "
-            "timeseries"
-    );
-    // GIL will be acquired if there is a string that is not pure ASCII/UTF-8
-    // In this case a PyObject will be allocated by convert::py_unicode_to_buffer
-    // If such a string is encountered in a column, then the GIL will be held until that whole column has
-    // been processed, on the assumption that if a column has one such string it will probably have many.
-    std::optional<ScopedGILLock> scoped_gil_lock;
-    const std::string_view column_name = on_.front();
-    std::vector<std::vector<size_t>> result;
-    result.resize(source_->num_rows);
-    const size_t source_field_position = field_index_for_matching_on_column(column_name, source_descriptor);
-    const Field& source_field = source_descriptor.field(source_field_position);
-    details::visit_type(source_field.type().data_type(), [&](auto source_field_dt) {
-        using SourceTDT = ScalarTagType<std::decay_t<decltype(source_field_dt)>>;
-        using SourceRawType = typename SourceTDT::DataTypeTag::raw_type;
-        const ColumnWithStrings target_column = std::get<ColumnWithStrings>(proc.get(ColumnName{column_name}));
-        details::visit_type(target_column.column_->type().data_type(), [&](auto target_field_dt) {
-            using TargetTDT = ScalarTagType<decltype(target_field_dt)>;
-            if constexpr (std::is_same_v<std::decay_t<SourceTDT>, std::decay_t<TargetTDT>>) {
-                using SourceType =
-                        std::conditional_t<is_sequence_type(SourceTDT::data_type()), PyObject* const, SourceRawType>;
-                auto target_values_to_rows = map_column_values_to_rows<TargetTDT>(target_column);
-                const size_t column_position_in_source_tensors =
-                        source_field_position - source_descriptor.index().field_count();
-                std::span source_data(
-                        static_cast<const SourceType*>(source_->field_tensors()[column_position_in_source_tensors].data(
-                        )),
-                        source_->num_rows
-                );
-                for (size_t source_row_idx = 0; source_row_idx < source_data.size(); ++source_row_idx) {
-                    auto source_value = source_data[source_row_idx];
-                    if constexpr (is_sequence_type(SourceTDT::data_type())) {
-                        if (is_py_none(source_value) || is_py_nan(source_value)) {
-                            result[source_row_idx] = target_values_to_rows[std::nullopt];
-                        } else {
-                            util::variant_match(
-                                    create_py_object_wrapper_or_error<SourceTDT::data_type()>(
-                                            source_value, scoped_gil_lock
-                                    ),
-                                    [&](convert::StringEncodingError&& err) {
-                                        err.row_index_in_slice_ = source_row_idx;
-                                        err.raise(column_name, 0);
-                                    },
-                                    [&](convert::PyStringWrapper&& wrapper) {
-                                        std::string_view wrapper_view{wrapper.buffer_, wrapper.length_};
-                                        result[source_row_idx] = target_values_to_rows[std::optional{wrapper_view}];
-                                    }
-                            );
-                        }
-                    } else {
-                        result[source_row_idx] = target_values_to_rows[source_value];
-                    }
-                }
-            } else {
-                schema::raise<ErrorCode::E_DESCRIPTOR_MISMATCH>(
-                        "Source type {} does not match target type {}. Dynamic schema is not implemented yet",
-                        SourceTDT::data_type(),
-                        TargetTDT::data_type()
-                );
-            }
-        });
-    });
-    return result;
-}
-
-std::pair<size_t, size_t> MergeUpdateClause::get_source_start_end(const ProcessingUnit& proc) const {
-    if (source_->has_index()) {
-        const std::span<const std::shared_ptr<AtomKey>> atom_keys = *proc.atom_keys_;
-        const auto source_row_range_it = source_start_end_for_row_range_.find(atom_keys.front()->time_range());
-        internal::check<ErrorCode::E_ASSERTION_FAILURE>(
-                source_row_range_it != source_start_end_for_row_range_.end(),
-                "Missing mapping between AtomKey timerange {} {} and source row start/end",
-                atom_keys.front()->time_range().first,
-                atom_keys.front()->time_range().second
-        );
-        return source_row_range_it->second;
-    } else {
-        return std::pair{0, source_->num_rows};
-    }
-}
-
-bool MergeUpdateClause::update_and_insert(
-        const std::span<const NativeTensor> source_tensors, const StreamDescriptor& source_descriptor,
-        const ProcessingUnit& proc, const std::span<const std::vector<size_t>> rows_to_update
-) const {
-    const std::span<const std::shared_ptr<SegmentInMemory>> target_segments = *proc.segments_;
-    const std::span<const std::shared_ptr<RowRange>> row_ranges = *proc.row_ranges_;
-    const std::span<const std::shared_ptr<ColRange>> col_ranges = *proc.col_ranges_;
-    const auto [source_row_start, source_row_end] = get_source_start_end(proc);
-    bool row_slice_changed = false;
-    const bool is_timeseries = source_->desc().index().type() == IndexDescriptor::Type::TIMESTAMP;
-    util::BitSet matched_target_rows(row_ranges.front()->diff());
-    // TODO: This can be inlined in the loop iterating over all columns to avoid iterating the source one more
-    // time. The loop structure makes it not intuitive. The performance cost must be evaluated. Monday: 10655963947
-    for (size_t source_row_idx = 0; source_row_idx < rows_to_update.size(); ++source_row_idx) {
-        for (const size_t target_row : rows_to_update[source_row_idx]) {
-            user_input::check<ErrorCode::E_INVALID_USER_ARGUMENT>(
-                    !matched_target_rows[target_row],
-                    "Multiple source rows match the same target row. This is not allowed as it's not clear which "
-                    "source row should be used for the update. Target row is {}, the second source row to match it is "
-                    "{}",
-                    target_row,
-                    source_row_idx + source_row_start
-            );
-            matched_target_rows[target_row] = true;
-        }
-    }
-    // Update one column at a time to increase cache coherency and to avoid calling visit_field for each row
-    // being updated
-    for (size_t segment_idx = 0; segment_idx < target_segments.size(); ++segment_idx) {
-        StringPool new_string_pool;
-        bool segment_contains_string_column = false;
-        SegmentInMemory& target_segment = *target_segments[segment_idx];
-        const size_t slice_size = target_segment.num_columns();
-        const int index_fields = source_descriptor.index().field_count();
-        for (size_t column_index_in_slice = index_fields; column_index_in_slice < slice_size; ++column_index_in_slice) {
-            const Field& target_field = target_segment.descriptor().field(column_index_in_slice);
-            details::visit_type(target_field.type().data_type(), [&]<typename DataTypeTag>(DataTypeTag) {
-                using RawType = DataTypeTag::raw_type;
-                using ScalarType = ScalarTagType<DataTypeTag>;
-                // All column slices start with the index. Subtract the index field count so that we don't count the
-                // index twice.
-                const size_t column_position_in_ts_descriptor =
-                        col_ranges[segment_idx]->start() + column_index_in_slice - index_fields;
-                // For pandas input the index NativeTensor is stored separately from the field NativeTensors.
-                // Subtract the index fields to get the correct position in the source descriptor
-                const size_t column_position_in_source = column_position_in_ts_descriptor - index_fields;
-                user_input::check<ErrorCode::E_INVALID_USER_ARGUMENT>(
-                        util::is_cstyle_array<RawType>(source_tensors[column_position_in_source]),
-                        "Fortran-style arrays are not supported by merge update yet. Column \"{}\" has data type "
-                        "{} of "
-                        "size {} bytes but the stride is {} bytes",
-                        target_field.name(),
-                        target_field.type(),
-                        sizeof(RawType),
-                        source_tensors[column_position_in_source].strides()[0]
-                );
-                Column& target_column = target_segment.column(column_index_in_slice);
-                const NativeTensor& source_tensor = source_tensors[column_position_in_source];
-                using SourceType =
-                        std::conditional_t<is_sequence_type(DataTypeTag::data_type), PyObject* const, RawType>;
-                const std::span source_data(
-                        static_cast<const SourceType*>(source_tensor.data()) + source_row_start,
-                        source_row_end - source_row_start
-                );
-                internal::check<ErrorCode::E_ASSERTION_FAILURE>(
-                        rows_to_update.size() == source_data.size(), "Mismatched source row sizes"
-                );
-                internal::check<ErrorCode::E_ASSERTION_FAILURE>(
-                        !rows_to_update.empty(), "There must be at least one source row inside the target row slice."
-                );
-                if constexpr (is_sequence_type(DataTypeTag::data_type)) {
-                    // String columns are always recreated from scratch regardless if an update or insert is
-                    // happening. This is done because the string pool must be updated. With data read from disk,
-                    // the map_ member of the pool is not populated. Which means that the mapping between a string
-                    // and offset in the pool is missing. To get it, we need to rebuild the pool anyway.
-                    segment_contains_string_column = true;
-                    const MergeUpdateStringColumnFlags column_update_flags{
-                            .is_timeseries = is_timeseries,
-                            .data_not_changed = on_set_.contains(target_field.name()) && strategy_.update_only()
-                    };
-                    row_slice_changed |= merge_update_string_column<ScalarType>(
-                            target_column,
-                            column_update_flags,
-                            rows_to_update,
-                            target_field,
-                            *row_ranges[segment_idx],
-                            target_segment.string_pool(),
-                            source_data,
-                            new_string_pool
-                    );
-                } else {
-                    if (is_update_only() && on_set_.contains(target_field.name())) {
-                        return;
-                    }
-                    ColumnData target_column_data = target_column.data();
-                    if (is_timeseries) {
-                        auto target_row_to_update_it = target_column_data.begin<ScalarType>();
-                        size_t target_row_to_update_idx = 0;
-                        for (size_t source_row_idx = 0; source_row_idx < source_data.size(); ++source_row_idx) {
-                            // TODO: Handle insert. Empty rows_to_update[source_row_idx] means that update must be done
-                            row_slice_changed |= !rows_to_update[source_row_idx].empty();
-                            for (const size_t target_row_idx : rows_to_update[source_row_idx]) {
-                                const size_t rows_to_skip = target_row_idx - target_row_to_update_idx;
-                                std::advance(target_row_to_update_it, rows_to_skip);
-                                *target_row_to_update_it = source_data[source_row_idx];
-                                target_row_to_update_idx = target_row_idx;
-                            }
-                        }
-                    } else {
-                        auto target = random_accessor<ScalarType>(&target_column_data);
-                        for (size_t source_row_idx = 0; source_row_idx < source_data.size(); ++source_row_idx) {
-                            // TODO: Handle insert. Empty rows_to_update[source_row_idx] means that update must be done
-                            row_slice_changed |= !rows_to_update[source_row_idx].empty();
-                            for (const size_t target_row_idx : rows_to_update[source_row_idx]) {
-                                target[target_row_idx] = source_data[source_row_idx];
-                            }
-                        }
-                    }
-                }
-            });
-        }
-        if (segment_contains_string_column) {
-            target_segment.set_string_pool(std::make_shared<StringPool>(std::move(new_string_pool)));
-        }
-    }
-    return row_slice_changed;
-}
-
-/// For each row of source that falls in the row slice in proc find all rows whose index matches the source index
-/// value. The matching rows will be sorted in increasing order. Since both source and target are timestamp indexed
-/// and ordered, only forward iteration on both source and target is needed and binary search can be used to check
-/// if a source index value exists in the target index. At the end some vectors in the output can be empty which
-/// means that that particular row in source did not match anything in the target. It is allowed for one row in
-/// target to be matched by multiple rows in source only if MergeUpdateClause::on_ is not empty. If
-/// MergeUpdateClause::on_ is not empty, there will be further filtering that might remove some matches. Otherwise,
-/// one row will be updated multiple times, which is not allowed.
-///
-/// Complexity: $$O(m * log_2(n) + n)$$ where: m is the count of source rows in the bounds of the segment, n is the
-/// number of target rows in the segment.
-std::vector<std::vector<size_t>> MergeUpdateClause::filter_index_match(
-        const Column& target_index, const std::span<const timestamp> source_index, const ProcessingUnit& proc
-) const {
-    using IndexType = ScalarTagType<DataTypeTag<DataType::NANOSECONDS_UTC64>>;
-    ColumnData target_index_column_data = target_index.data();
-    const auto target_index_accessor = random_accessor<IndexType>(&target_index_column_data);
-    const auto [source_row_start, source_row_end] = get_source_start_end(proc);
-    const size_t last_target_row_to_consider = [&] {
-        auto row_indexes = ranges::iota_view{position_t{0}, target_index.row_count()};
-        const auto upper_bound = ranges::upper_bound(
-                row_indexes,
-                source_index.back(),
-                [&](const timestamp value, const int64_t row_idx) { return value < target_index_accessor.at(row_idx); }
-        );
-        return upper_bound == row_indexes.end() ? target_index.row_count() : *upper_bound;
-    }();
-    size_t source_row = source_row_start;
-    const size_t source_rows_in_row_slice = source_row_end - source_row;
-    std::vector<std::vector<size_t>> matched_rows(source_rows_in_row_slice);
-    size_t target_row = 0;
-    // This loop can be inverted so that if source_row_end - source_row_start is > len(target) the complexity becomes
-    // O(n * log_2(m)) where: m is the count of source rows in the bounds of the segment, n is the number of target rows
-    // in the segment.
-    while (target_row < last_target_row_to_consider && source_row < source_row_end) {
-        const timestamp source_ts = source_index[source_row];
-        // TODO: Profile performance and try different optimizations. See Monday 10655963947
-        auto row_index_view = std::ranges::iota_view{target_row, last_target_row_to_consider};
-        auto target_match_it =
-                ranges::lower_bound(row_index_view, source_ts, [&](const size_t row_idx, const timestamp source) {
-                    return target_index_accessor.at(row_idx) < source;
-                });
-        if (target_match_it == row_index_view.end()) {
-            break;
-        }
-        target_row = *target_match_it;
-        while (target_row < last_target_row_to_consider && target_index_accessor.at(target_row) == source_ts) {
-            matched_rows[source_row - source_row_start].push_back(target_row);
-            ++target_row;
-        }
-        // Optimizes the case of repeated index values. All matched rows corresponding to a particular index value must
-        // be the same, so just copy the matched rows in case index values are repeated.
-        while (++source_row < source_row_end && source_index[source_row] == source_ts) {
-            const size_t source_row_offset = source_row - source_row_start;
-            matched_rows[source_row_offset] = matched_rows[source_row_offset - 1];
-        }
-    }
-    return matched_rows;
-}
-
-/// Complexity: $$O(c * n * m)$$ m is the count of source rows in the bounds of the segment, n is the  number of target
-/// rows in the segment. c is the number of columns in MergeUpdateClause::on_. It can be reached if the data in source
-/// and target is the same up to the very last column in MergeUpdateClause::on_.
-std::vector<std::vector<size_t>> MergeUpdateClause::filter_on_additional_columns_match(
-        const StreamDescriptor& source_descriptor, const StreamDescriptor& target_descriptor,
-        const std::span<const NativeTensor> source_tensors, ProcessingUnit& proc,
-        std::optional<std::vector<std::vector<size_t>>>&& index_match
-) const {
-    ranges::subrange on = on_;
-    std::vector<std::vector<size_t>> matched_rows;
-    const IndexDescriptor::Type source_index_type = source_descriptor.index().type();
-    const IndexDescriptor::Type target_index_type = target_descriptor.index().type();
-    if (index_match) {
-        user_input::check<ErrorCode::E_INVALID_USER_ARGUMENT>(
-                (source_index_type == target_index_type) && (source_index_type == IndexDescriptor::Type::TIMESTAMP),
-                "Source and target index types must both be TIMESTAMP. Source: {}, target: {}",
-                source_index_type,
-                target_index_type
-        );
-        matched_rows = std::move(*index_match);
-    } else {
-        user_input::check<ErrorCode::E_INVALID_USER_ARGUMENT>(
-                (source_index_type == target_index_type) && (source_index_type == IndexDescriptor::Type::ROWCOUNT),
-                "Source and target index types to be both ROWCOUNT. Source: {}, target: {}",
-                source_index_type,
-                target_index_type
-        );
-        matched_rows = initialize_rows_to_update_for_rowrange_indexed_data(proc, source_descriptor);
-        on = on.next();
-    }
-    if (on.empty()) {
-        return matched_rows;
-    }
-    const auto [source_row_start, source_row_end] = get_source_start_end(proc);
-    // GIL will be acquired if there is a string that is not pure ASCII/UTF-8
-    // In this case a PyObject will be allocated by convert::py_unicode_to_buffer
-    // If such a string is encountered in a column, then the GIL will be held until that whole column has
-    // been processed, on the assumption that if a column has one such string it will probably have many.
-    std::optional<ScopedGILLock> scoped_gil_lock;
-    for (std::string_view column_name : on) {
-        const size_t source_field_position = field_index_for_matching_on_column(column_name, source_descriptor);
-        const Field& source_field = source_descriptor.field(source_field_position);
-        details::visit_type(source_field.type().data_type(), [&]<typename SourceDataTypeTag>(SourceDataTypeTag) {
-            using SourceTDT = ScalarTagType<SourceDataTypeTag>;
-            using SourceType = std::conditional_t<
-                    is_sequence_type(SourceDataTypeTag::data_type),
-                    PyObject* const*,
-                    const typename SourceDataTypeTag::raw_type*>;
-            const size_t column_position_in_source_tensors =
-                    source_field_position - source_descriptor.index().field_count();
-            std::span source_data(
-                    static_cast<SourceType>(source_tensors[column_position_in_source_tensors].data()) +
-                            source_row_start,
-                    source_row_end - source_row_start
-            );
-            // TODO: For dynamic schema the two fields might have different indexes
-            const size_t target_field_position = source_field_position;
-            const ColumnWithStrings target_column = std::get<ColumnWithStrings>(
-                    proc.get(ColumnName{target_descriptor.field(target_field_position).name()})
-            );
-            ColumnData target_data = target_column.column_->data();
-            details::visit_type(
-                    target_column.column_->type().data_type(),
-                    [&]<typename TargetDataTypeTag>(TargetDataTypeTag) {
-                        using TargetTDT = ScalarTagType<TargetDataTypeTag>;
-                        using TargetRawType = TargetDataTypeTag::raw_type;
-                        // TODO: Relax for dynamic schema
-                        if constexpr (std::same_as<std::decay_t<SourceDataTypeTag>, std::decay_t<TargetDataTypeTag>>) {
-                            auto target_accessor = random_accessor<TargetTDT>(&target_data);
-                            for (size_t source_row_idx = 0; source_row_idx < matched_rows.size(); ++source_row_idx) {
-                                std::erase_if(matched_rows[source_row_idx], [&](const size_t target_row_idx) {
-                                    const TargetRawType target_value = target_accessor.at(target_row_idx);
-                                    const auto& source_value = source_data[source_row_idx];
-                                    auto are_values_equal = are_merge_values_matching<SourceTDT, TargetTDT>(
-                                            source_value, target_value, *target_column.string_pool_, scoped_gil_lock
-                                    );
-                                    if constexpr (is_sequence_type(TargetDataTypeTag::data_type)) {
-                                        return util::variant_match(
-                                                are_values_equal,
-                                                [](const bool equal) { return !equal; },
-                                                [&](convert::StringEncodingError& err) -> bool {
-                                                    err.row_index_in_slice_ = source_row_idx;
-                                                    err.raise(column_name, source_row_start);
-                                                }
-                                        );
-                                    } else {
-                                        return !std::get<bool>(are_values_equal);
-                                    }
-                                });
-                            }
-                        } else {
-                            internal::raise<ErrorCode::E_ASSERTION_FAILURE>(
-                                    "Target column \"{}\" has unexpected type {}. Source type: {}",
-                                    column_name,
-                                    TargetDataTypeTag::data_type,
-                                    SourceDataTypeTag::data_type
-                            );
-                        }
-                    }
-            );
-        });
-    }
-    return matched_rows;
-}
-
-const ClauseInfo& MergeUpdateClause::clause_info() const { return clause_info_; }
-
-void MergeUpdateClause::set_processing_config(const ProcessingConfig&) {}
-
-void MergeUpdateClause::set_component_manager(std::shared_ptr<ComponentManager> component_manager) {
-    component_manager_ = std::move(component_manager);
-}
-
-OutputSchema MergeUpdateClause::modify_schema(OutputSchema&& output_schema) const {
-    schema::check<ErrorCode::E_DESCRIPTOR_MISMATCH>(
-            columns_match(output_schema.stream_descriptor(), source_->desc()),
-            "Cannot perform merge update when the source and target schema are not the same.\nSource schema: "
-            "{}\nTarget schema: {}",
-            source_->desc(),
-            output_schema.stream_descriptor()
-    );
-    return output_schema;
-}
-
-OutputSchema MergeUpdateClause::join_schemas(std::vector<OutputSchema>&&) const {
-    util::raise_rte("MergeUpdateClause::join_schemas should never be called");
-}
-
-std::string MergeUpdateClause::to_string() const { return "MERGE_UPDATE"; }
-
-bool MergeUpdateClause::is_update_only() const {
-    return strategy_ == MergeStrategy{MergeAction::UPDATE, MergeAction::DO_NOTHING};
-}
-
-size_t MergeUpdateClause::field_index_for_matching_on_column(std::string_view name, const StreamDescriptor& descriptor)
-        const {
-    // In case of unnamed index columns we set the fake_name property of the metadata to true and assign the name
-    // "index" to the column. In this specific case the user must be able to match on a column named "index".
-    user_input::check<ErrorCode::E_INVALID_USER_ARGUMENT>(
-            descriptor.index().type() != IndexDescriptor::Type::TIMESTAMP ||
-                    (descriptor.field(0).name() != name || (fake_index_name_ && descriptor.field(0).name() == name)),
-            "The \"on\" parameter must not contain the datetime index column \"{}\". Note that for date time indexed "
-            "column the index column is always used for matching. Descriptor: {}",
-            name,
-            descriptor
-    );
-    const boost::regex repetition_regex{fmt::format("^__col_{}__\\d+$", name)};
-    const std::string multiindex_mangled_name = stream::mangled_name(name);
-    std::ranges::subrange data_fields{
-            std::next(descriptor.fields().begin(), descriptor.index().field_count()), descriptor.fields().end()
-    };
-    const auto is_name_or_mangled_name = [&](const Field& field) {
-        return field.name() == name || field.name() == multiindex_mangled_name ||
-               boost::regex_match(field.name().begin(), field.name().end(), repetition_regex);
-    };
-    const auto field_it = std::ranges::find_if(data_fields, is_name_or_mangled_name);
-    user_input::check<ErrorCode::E_COLUMN_NOT_FOUND>(
-            field_it != data_fields.end(),
-            "Column \"{}\" specified in the 'on' parameter does not exist. Descriptor: {}",
-            name,
-            descriptor
-    );
-    const auto repetition_it = std::ranges::find_if(
-            data_fields.next(std::distance(data_fields.begin(), field_it) + 1), is_name_or_mangled_name
-    );
-    user_input::check<ErrorCode::E_DUPLICATE_COLUMN>(
-            repetition_it == data_fields.end(),
-            "Column \"{}\" specified in the 'on' appears more than once in the dataframe. Descriptor: {}",
-            name,
-            descriptor
-    );
-    return std::distance(descriptor.fields().begin(), field_it);
-}
 
 CompactDataClause::CompactDataClause(uint64_t rows_per_segment) : rows_per_segment_(rows_per_segment) {
     // Magic range of +-33% chosen so that:
