+++ cpp/arcticdb/version/versioned_engine.hpp
@@ -9,8 +9,8 @@
 #pragma once
 
 #include <arcticdb/version/version_map.hpp>
-#include <arcticdb/async/async_store.hpp>
 #include <arcticdb/version/symbol_list.hpp>
+#include <arcticdb/storage/common.hpp>
 #include <arcticdb/entity/protobufs.hpp>
 #include <arcticdb/entity/stage_result.hpp>
 #include <arcticdb/pipeline/write_options.hpp>
